# Adjusted the file loading # Joined some methods in the PackerFileLoader class # Added a check on the PacketObjectManager class to verify if we are trying to add a file that has a bigger size then our maximum fragment capacity # Updated the gitignore

+++ Packet/PacketFragment.h
@@ -61,7 +61,7 @@ class PacketFragment
 public:
 
 	// The fragment file type
-	typedef uint32_t FileIdentifier;
+	typedef uint64_t FileIdentifier;
 
 	// The section metadata type
 	struct SectionMetadata
